Move worker pieces that are only need once to init.pp
This commit moves all the pieces from worker.pp which are used by all
the worker process into init.pp. This was not done before because the
package definitions would conflict with those in logstash.o.o's
manifest. However with the if not defined checks we can have them in
init.pp, so let's move everything in there.

Change-Id: I5048604ce64bb52f463397c6a37b4a19a0bab8bb