Add refreshonly to 'upgrade_subunit2sql_db'
This commit adds refreshonly to the 'upgrade_subunit2sql_db' exec.
The problem without this set is that subscribing to the Pip package
generates an event on each puppet refresh, which causes the db-manage
utility to run a migration update every 15 min, which is far from
ideal. Setting refreshonly means that the command is only run when
the resource changes, in this case when a new package version is
installed, which is the desired behavior.

Change-Id: Ib77d68f19edbf48a8d8cf4e757dbfb6463f1c6bb