Remove subunit v1 to v2 conversion step
We stopped storing subunit v1 on the log servers a long time ago, but
there wasn't a pressing need to stop doing the conversion stage because
a v2 stream will just pass through subunit-1to2. However, the subprocess
command used for that step was leaking FDs which is causing issues. So
this commit fixes the leak by just removing the conversion stage since
it no longer has a purpose anyway.

Change-Id: I04c6120c2c98dcd17eb21c908919450b11934100