Ensure we close the subunit_v2 stream file object
This subunit_v2 stream file object (or file like object) that gets
eventually passed into the event processor is never closed. The
read_subunit class only does a flush() on the object when it's done
with it. This commit adds a close() after all the processing is done
just to ensure we don't leak in the future.

Change-Id: I2975d30f772a1fe60489b7d6a9f4ae50d0d87e7f