Add more debug logging for closed file issues
This commit adds additional debug logging to the subunit gearman
worker to try and identify the source of the closed file bug we're
hitting in production. This adds logging of the memory address where
the cStringIO.StringIO objects before and after being pushed on the
queue. It also adds a distinct message for the object if it's closed
before or after being on the queue.

Change-Id: I4a4f064e8885f8f9c8fc2974c0f71d837f41454e