Remove inaccurate comment
This commit removes an inaccurate comment that got copy and pasted
over when the script was originally created. The StdOut processor from
the logstash gearman client only works in foreground mode (for obvious
reasons) and since it was the simplest example that was used as a basis
for the subunit processor. However, the subunit processor doesn't share
this limitation and can (and has been) run as a daemon in the
background.

Change-Id: I332523567d16b0994b06e278b68ffe8bcb7d9bfa