Stop using a seperate retreiver thread
Since we recently switched from having 2 threads, 1 to retreive
subunit and 1 to process it, to having a single thread do both there
isn't any reason to launch to run that in a separate thread anymore.
This commit just removes the use of threading and runs everything in
the single process.

Change-Id: I5205fc73178b7d5a4bbee61e68b16b63499f5dd8