Switch to storing run wall time in runs database
This commit uses an option added in the subunit2sql 1.8.0 release to
set the run_time column in the runs table to be the wall time for the
subunit stream, in other words the stop time of the last test - the
start time of the first test. By default subunit2sql stores the sum of
all the individual test run times as the value for the run_time column
which is of questionable value for doing analysis. By switching what we
store for this value in the infra DB hopefully this data will be of more
use to people.

Change-Id: Ief4cfa91f9661444b3680b428da0a3c4ca5dedd0