Config is a dict so use get not getattr
The config file is a yaml and parsed into a dict. However there was a
mistake in the recent patch that treated it as a ConfigParser object and
used getattr to access the mqtt options. This doesn't work and needs to
just be a get to access the key from the dict. This commit does that so
now it should work without crashing.

Change-Id: Ia3613adb6a037c9d406e6e33361d6d5ee826c9cd