Ensure that build_names are unique per project
With the migration to zuul v3 project names for common jobs, like unit
tests, no longer are unique per project. This causes issues for
openstack-health and subunit2sql because we can longer easily filter a
unit test job between projects. This commit adds a function to ensure
that the project name is in the build_name in the metadata. (assuming
devstack or tempest is not in the job_name) This way we can still filter
based on the job name per project.

Change-Id: I5231105af975d987d5f5e5f77c6ac038e10cd832